X.509 OBS auth: Only prolong proxy if needed, fix SELinux issue.
This changes proxy renewal to only recreate the proxy (72 hours lifetime)
if the remaining lifetime drops below 24 hours.
In addition, it is copied to /var/run/condor/proxy which is now used
in the pilot JDLs for HTCondor local batch systems.

This significantly reduces the number of filetransfers into
HTCondor drones, and fixes an SELinux issue
(shadow needs to access the proxy).